requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
feedparser==6.0.11
sentence-transformers==2.3.1
gliner==0.1.5
numpy==1.26.3
scikit-learn==1.4.0
bertopic==0.16.0
pydantic==2.6.0
torch
requests==2.31.0
google-generativeai==0.3.0
python-dotenv==1.0.0
bertopic==0.16.0
fastapi==0.109.0
faiss-cpu
feedparser==6.0.11
google-generativeai==0.3.0
numpy==1.26.3
pydantic==2.6.0
python-dateutil
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.4.0
sentence-transformers==2.3.1
torch
tqdm
uvicorn==0.27.0
google.generativeai
googlenewsdecoder
google-search-results
google-cloud-translate==2.0.1
langdetect==1.0.9
uvicorn
tqdm
faiss-cpu
aiohttp
aiosqlite
fast-langdetect
indictrans2
indictranstoolkit
